Merged PR 1109: Fix for media viewer zoom issue
This pr contains a fix for image being visible above topbar in zoom mode.
I propose to hide the topbar in zoom mode because the library we are using (https://www.npmjs.com/package/@likashefqet/react-native-image-zoom) has inbuilt transparent background color for the statusbar. So the image is visible under the status bar. UI wise, it looks better to remove topbar rather than remove zoomed image in that space. Earlier it was thought that this was SafeAreaView issue but it's not. It's how the library is set up.