Merged PR 1460: Patched issues with sharing links or text into the app
While sharing into text or weblinks into the app, the app would enter a weird state where it assumed that files were being shared when in fact they were not.

To resolve this we first add the constraint that if a user is sharing such a content type, they can only share one such message. This holds true across other apps as well. We then use this constraint to implement a test for whether we're sharing a file that actually works.