Merged PR 1475: Edit message feature for direct and group chat
This feature allows the user to edit a message on a direct and a group chat.
User can edit a message multiple number of times.
There are safeguards to not let the user edit a received message.
This pr also has a new sender and a new receiver for edited messages.
Edit message is currently supported for text, images and videos.

Related work items: #1161, #1162