Merged PR 1539: Linked Superport card design in FolderChats
Prev we used to show 'Superports' card at the middle in placeholder and it simply redirects user to superport creation flow with redirected folder as linked folder for superport.
And even in cases when the folder is connected to some supeport which shows 'folder is linked to 1 superport' same redirection functionality was there.

Now the 'Linked supeport' card will always be at the top, if linked superport length is more than 0 then it will redirect you to the superports screen with filter as that folder, else redirect to creation flow with pre selected linked folder.

New vs Old Screen:
![Screenshot 2024-11-12 at 6.51.41 PM.png](https://dev.azure.com/numberless-india/dc5ba45b-af2e-45e7-b8c3-90a606b4bf3f/_apis/git/repositories/4075e82c-5db8-4d88-b417-30b6261033b9/pullRequests/1539/attachments/Screenshot%202024-11-12%20at%206.51.41%E2%80%AFPM.png)