Merged PR 1581: Folder chip background color miss match
at most places where we are using bold and light accent colors, we are mapping the color object.
so i found one extra color in light accent color, hence the mapping was bad and the bg color for the folder chip on superport was different.

So in this pr, the extra 'grey' color was only being used at one place, which was for the 'favourite' permission icon.
so keeping that an exception i kept the color in primary object instead of accent colors, cause it is not an accent color and doesn't belong there.

issue:
the bg color should be light teal here.
![Screenshot 2024-11-20 at 9.12.05 PM.png](https://dev.azure.com/numberless-india/dc5ba45b-af2e-45e7-b8c3-90a606b4bf3f/_apis/git/repositories/4075e82c-5db8-4d88-b417-30b6261033b9/pullRequests/1581/attachments/Screenshot%202024-11-20%20at%209.12.05%E2%80%AFPM.png)

an extra color grey was added in low accent color but not dark, so added exception condition in get folder permission icon for grey